restructure into lib and data dirs, hide indexer
