+++ lib/chunkWorker.js
