+++ lib/generateIndex.js
