+++ lib/lfqueue.js
