+++ lib/zcbuf.js
